+++ test/int/test_runtime_configuration.py
@@ -40,19 +40,19 @@ def test_set_via_alter_system(cluster: Cluster):
     # stage before creating DML from ir node
     with pytest.raises(
         TarantoolError,
-        match="""invalid value for 'iproto_net_msg_max' expected unsigned, got integer""",
+        match="""invalid value for 'iproto_net_msg_max': expected unsigned, got integer""",
     ):
         instance.sql("ALTER SYSTEM SET iproto_net_msg_max = -1")
 
     with pytest.raises(
         TarantoolError,
-        match="""invalid value for 'memtx_checkpoint_interval' expected unsigned, got decimal""",
+        match="""invalid value for 'memtx_checkpoint_interval': expected unsigned, got decimal""",
     ):
         instance.sql("ALTER SYSTEM SET memtx_checkpoint_interval = -1.0")
 
     with pytest.raises(
         TarantoolError,
-        match="""invalid value for 'memtx_checkpoint_count' expected unsigned, got integer""",
+        match="""invalid value for 'memtx_checkpoint_count': value must be between 1 and 2147483647""",
     ):
         instance.sql("ALTER SYSTEM SET memtx_checkpoint_count = -1")
 
@@ -260,3 +260,320 @@ def test_cache_capacity(cluster: Cluster):
     # because of LRU
     cache_info = i1.eval("return box.info.sql()")
     assert cache_info["cache"]["size"] == 2442
+
+
+def test_alter_system_iproto_net_msg_max(cluster: Cluster):
+    instance = cluster.add_instance()
+    iproto_net_msg_max = "iproto_net_msg_max"
+
+    # normal value
+    instance.sql(f"ALTER SYSTEM SET {iproto_net_msg_max} = 420")
+
+    bad_value = 0
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{iproto_net_msg_max}': value must be between 2 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {iproto_net_msg_max} = {bad_value}")
+
+    bad_value = 1
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{iproto_net_msg_max}': value must be between 2 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {iproto_net_msg_max} = {bad_value}")
+
+    bad_value = 2147483648
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{iproto_net_msg_max}': value must be between 2 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {iproto_net_msg_max} = {bad_value}")
+
+    bad_value = False
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{iproto_net_msg_max}': expected unsigned, got boolean""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {iproto_net_msg_max} = {bad_value}")
+
+    str_value = "str"
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{iproto_net_msg_max}': expected unsigned, got string""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {iproto_net_msg_max} = '{str_value}'")
+
+
+def test_alter_system_memtx_checkpoint_count(cluster: Cluster):
+    instance = cluster.add_instance()
+    memtx_checkpoint_count = "memtx_checkpoint_count"
+
+    # normal value
+    instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_count} = 4")
+
+    bad_value = -1
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_count}': value must be between 1 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_count} = {bad_value}")
+
+    bad_value = 0
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_count}': value must be between 1 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_count} = {bad_value}")
+
+    bad_value = 2147483648
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_count}': value must be between 1 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_count} = {bad_value}")
+
+    bad_value = True
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_count}': expected integer, got boolean""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_count} = {bad_value}")
+
+    str_value = "str"
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_count}': expected integer, got string""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_count} = '{str_value}'")
+
+
+def test_alter_system_memtx_checkpoint_interval(cluster: Cluster):
+    instance = cluster.add_instance()
+    memtx_checkpoint_interval = "memtx_checkpoint_interval"
+
+    # normal value
+    instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_interval} = 3600")
+
+    bad_value = -1
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_interval}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_interval} = {bad_value}")
+
+    bad_value = False
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{memtx_checkpoint_interval}': expected unsigned, got boolean""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {memtx_checkpoint_interval} = {bad_value}")
+
+
+def test_alter_system_sql_storage_cache_size_max(cluster: Cluster):
+    instance = cluster.add_instance()
+    sql_storage_cache_size_max = "sql_storage_cache_size_max"
+
+    # normal vlaue
+    instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_size_max} = 15000")
+
+    bad_value = 0
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_size_max}': value must be between 1 and 2147483647""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_size_max} = {bad_value}")
+
+    bad_value = 1
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_size_max}': value must be greater than the current cache size 851""",
+    ):
+        instance.sql("SELECT 1")
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_size_max} = {bad_value}")
+
+    bad_value = -1
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_size_max}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_size_max} = {bad_value}")
+
+    bad_value = True
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_size_max}': expected unsigned, got boolean""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_size_max} = {bad_value}")
+
+
+def test_alter_system_sql_storage_cache_count_max(cluster: Cluster):
+    instance = cluster.add_instance()
+    sql_storage_cache_count_max = "sql_storage_cache_count_max"
+
+    instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_count_max} = 15000")
+
+    bad_value = 0
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_count_max}': value must be between 1 and 18446744073709551615""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_count_max} = {bad_value}")
+
+    bad_value = -1
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_count_max}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_count_max} = {bad_value}")
+
+    bad_value = False
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_storage_cache_count_max}': expected unsigned, got boolean""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_storage_cache_count_max} = {bad_value}")
+
+
+def test_alter_system_motion_row_param(cluster: Cluster):
+    instance = cluster.add_instance(init_replication_factor=2)
+    i2 = cluster.add_instance()
+    sql_motion_row_max = "sql_motion_row_max"
+
+    db_config = instance.sql(f"""
+                SELECT key, value FROM _pico_db_config
+                WHERE key='{sql_motion_row_max}'
+                """)
+    assert db_config == [["sql_motion_row_max", 5000]]
+
+    # limit the parameter
+    limit = 25
+    instance.sql(f"ALTER SYSTEM SET {sql_motion_row_max} TO {limit}")
+    assert instance.sql(f"SELECT key, value FROM _pico_db_config WHERE key = '{sql_motion_row_max}'") == [
+        ["sql_motion_row_max", limit]
+    ]
+
+    i2.sql("CREATE TABLE t1(id INT PRIMARY KEY) DISTRIBUTED BY (id)")
+    i2.sql("CREATE TABLE t2(id INT PRIMARY KEY) DISTRIBUTED BY (id)")
+    size = limit * 2
+    for x in range(size):
+        i2.sql(f"INSERT INTO t1 VALUES ({x})")
+        i2.sql(f"INSERT INTO t2 VALUES ({x})")
+
+    with pytest.raises(TarantoolError, match="Exceeded maximum number of rows"):
+        instance.sql("SELECT * FROM t1 JOIN t2 ON true")
+
+    # set to 0 to disable limit and verify the same query now succeeds
+    limit = 0
+    instance.sql(f"ALTER SYSTEM SET {sql_motion_row_max} TO {limit}")
+    assert instance.sql(f"SELECT key, value FROM _pico_db_config WHERE key = '{sql_motion_row_max}'") == [
+        ["sql_motion_row_max", limit]
+    ]
+
+    row_count = instance.sql("SELECT COUNT(*) FROM t1 JOIN t2 ON true")
+    # 50 * 50 = 2500 rows total
+    assert row_count[0][0] == 2500
+
+    # bad values for parameters shouldn't pass validation
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_motion_row_max}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_motion_row_max} = -1")
+
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_motion_row_max}': expected unsigned, got string""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_motion_row_max} = 'x'")
+
+
+def test_alter_system_vbde_opcode_param(cluster: Cluster):
+    instance = cluster.add_instance(init_replication_factor=2)
+    sql_vdbe_opcode_max = "sql_vdbe_opcode_max"
+
+    db_config = instance.sql(f"""
+                SELECT key, value FROM _pico_db_config
+                WHERE key='{sql_vdbe_opcode_max}'
+                """)
+    assert db_config == [["sql_vdbe_opcode_max", 45000]]
+
+    # limit the parameter
+    limit = 100
+    instance.sql(f"ALTER SYSTEM SET {sql_vdbe_opcode_max} TO {limit}")
+    assert instance.sql(f"SELECT key, value FROM _pico_db_config WHERE key = '{sql_vdbe_opcode_max}'") == [
+        ["sql_vdbe_opcode_max", limit]
+    ]
+
+    with pytest.raises(TarantoolError, match=f"Reached a limit on max executed vdbe opcodes. Limit: {limit}"):
+        instance.sql("SELECT * FROM _pico_db_config")
+
+    # set 0 to disable limits
+    limit = 0
+    instance.sql(f"ALTER SYSTEM SET {sql_vdbe_opcode_max} TO {limit}")
+    row_count = instance.sql("SELECT * FROM _pico_db_config")
+    assert row_count != 0
+
+    # bad values for parameters shouldn't pass validation
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_vdbe_opcode_max}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_vdbe_opcode_max} = -1")
+
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{sql_vdbe_opcode_max}': expected unsigned, got string""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {sql_vdbe_opcode_max} = 'x'")
+
+
+def test_alter_system_pg_params(cluster: Cluster):
+    instance = cluster.add_instance()
+    pg_portal_max = "pg_portal_max"
+    pg_statement_max = "pg_statement_max"
+
+    # default values
+    db_config = instance.sql(f"""
+                SELECT key, value FROM _pico_db_config
+                WHERE key IN ('{pg_portal_max}', '{pg_statement_max}')
+                ORDER BY key
+                """)
+    assert db_config == [["pg_portal_max", 1024], ["pg_statement_max", 1024]]
+
+    # set 0 to disable limits
+    instance.sql(f"ALTER SYSTEM SET {pg_portal_max} TO 0")
+    instance.sql(f"ALTER SYSTEM SET {pg_statement_max} TO 0")
+
+    # parameters values changed
+    db_config = instance.sql(f"""
+                SELECT key, value FROM _pico_db_config
+                WHERE key IN ('{pg_portal_max}', '{pg_statement_max}')
+                ORDER BY key
+                """)
+    assert db_config == [["pg_portal_max", 0], ["pg_statement_max", 0]]
+
+    # bad values for parameters shouldn't pass validation
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{pg_portal_max}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {pg_portal_max} = -1")
+
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{pg_portal_max}': expected unsigned, got string""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {pg_portal_max} = 'x'")
+
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{pg_statement_max}': expected unsigned, got integer""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {pg_statement_max} = -1")
+
+    with pytest.raises(
+        TarantoolError,
+        match=f"""invalid value for '{pg_statement_max}': expected unsigned, got string""",
+    ):
+        instance.sql(f"ALTER SYSTEM SET {pg_statement_max} = 'y'")